LayoutTests/media/modern-media-controls/media-controller/media-controller-single-container-expected.txt

Testing the MediaController has a single container for captions and media controls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.childElementCount is 2
PASS shadowRoot.firstElementChild.localName is "style"
PASS shadowRoot.lastElementChild.localName is "div"
PASS shadowRoot.lastElementChild.className is "media-controls-container"
PASS shadowRoot.lastElementChild.childElementCount is 2
PASS shadowRoot.lastElementChild.firstElementChild.localName is "div"
PASS shadowRoot.lastElementChild.firstElementChild.getAttribute('pseudo') is "-webkit-media-text-track-container"
PASS shadowRoot.lastElementChild.lastElementChild.localName is "div"
PASS shadowRoot.lastElementChild.lastElementChild.classList.contains('media-controls') is true
PASS successfullyParsed is true

TEST COMPLETE
